% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{data_analysis}
\alias{data_analysis}
\title{Emperical ranks probbaility of the test given the effect size}
\usage{
data_analysis(alpha, pvalue, filter, N_current, N_prior, tail, max.group,
  effectType = c("continuous", "binary"))
}
\arguments{
\item{alpha}{Nmeric, significance level of the hypothesis test}

\item{pvalue}{a vector of pvalues of the test statistics}

\item{filter}{a vector of filter statistics}

\item{N_current}{Integer, number of observations in the current data}

\item{N_prior}{Integer, number of observations in the prior data}

\item{tail}{right-tailed or two-tailed hypothesis test.
default is right-tailed test.}

\item{max.group}{maximum number of groups to be used to split the p-values,
default is five. Note that, it is better to keep approximately 1000 p-values
per group.}

\item{effectType}{Character of type ("binary","continuous") of effect sizes}
}
\value{
\code{rejections} A numeric vector of the number of rejected test of
the different methods.
}
\description{
Emperical comnputation of the ranks probability of a test being
higher than any other test given the effect size from the external information.
}
\details{
Perform data analysis for the different methods such as proposed,
bonferroni, Benjamini and Hoghburgh, IHW, and Dorbibian methods.
}
\examples{

# generating data (known in practice)
set.seed(123)
m = 10000
X = runif(m, min = 0, max = 2.5)               # covariate
H = rbinom(length(X), size = 1, prob = 0.1)   # hypothesis true or false
Z = rnorm(length(X), mean = H * X)            # Z-score
p = 1 - pnorm(Z)
rejections <- data_analysis(alpha = .1, pvalue = p, filter = X, N_current = m,
         N_prior = m, tail = 2, max.group = 10, effectType = "continuous")

}
\author{
Mohamad S. Hasan, shakilmohamad7@gmail.com
}
