% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwerPowerFdrPower_emp.R
\name{fwerPowerFdrPower_emp}
\alias{fwerPowerFdrPower_emp}
\title{Simulate FWER, POWER, FDR, and POWER}
\usage{
fwerPowerFdrPower_emp(i, simu, m, null, corr = 0, cv = 0, alpha = 0.05,
  groupSize = 100L, max.group = 5L, filterEffectVec,
  effectType = c("continuous", "binary"))
}
\arguments{
\item{i}{i-th filter effect}

\item{simu}{number of replications}

\item{m}{number of hypothesis test}

\item{null}{proportion of the true null hypothesis}

\item{corr}{correlation between test statistics}

\item{cv}{determine whether the test mean effect and the filter mean effects
are the same}

\item{alpha}{significance threshold}

\item{groupSize}{number of test statistics per group}

\item{max.group}{maximum number of p-value groups to be used, minimum is 5}

\item{filterEffectVec}{a vector of different effect size}

\item{effectType}{type of effect sizes, c("continuous", "binary")}
}
\value{
a matrix of 16 rows containing information about FWER, POWER, FDR,
and POWER (4 rows for each)
}
\description{
This function simulate Family Wise Error Rate (FWER) and
corresponding Power, and False Discovery Rate (FDR) and the corresponding
Power for different effect sizes
}
\details{
This function empirically simulate Family Wise Error Rate (FWER) and
corresponding Power, and False Discovery Rate (FDR) and the corresponding
Power for the different effect sizes
}
\examples{
# vector of effect sizes
effectVec <- c(1, 1.5, 2)
simuVal = 2
FwerPowerFdrPower <- sapply(1:length(effectVec), fwerPowerFdrPower_emp,
             simu = simuVal, m = 10000, null = .5, corr = 0,
             cv = 0, alpha = .05, groupSize = 100, max.group = 5,
             filterEffectVec = effectVec, effectType = "continuous")

}
\references{
Hasan and Schliekelman (2017)
}
\seealso{
\code{\link{fwerPowerFdrPower}}
}
\author{
Mohamad S. Hasan, \email{mshasan@uga.edu}
}
