% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_plots2.R
\name{nice_plots2}
\alias{nice_plots2}
\title{Funciton to plot nice plots}
\usage{
nice_plots2(x_vec, y_matrix, x_name = NULL, y_names = NULL, xlab = NULL,
  ylab = NULL, title = NULL)
}
\arguments{
\item{x_vec}{A numeric vector of x-axis values}

\item{y_matrix}{A numeric matrix of y-axis values}

\item{x_name}{Character, name of the x variable}

\item{y_names}{Character vector, name of the y variables}

\item{xlab}{Character, lebel of x-axis}

\item{ylab}{Character, label of y-axis}

\item{title}{Character, title of the plot}
}
\value{
\code{Data}
A plot of multiple curves
}
\description{
This function takes a vector of xvalues and a vector or matrix of yvalues
then plot
}
\details{
This function is desigend to plot nice curves
}
\examples{
x_vec <- 1:10
y_matrix = cbind(sort(runif(10)), sort(rnorm(10)))
plt <- nice_plots2(x_vec = x_vec, y_matrix = y_matrix, x_name = "groups",
             y_names = c("uniform", "normal"),
             xlab = "groups", ylab = "Distribution", title = "uniform vs. normal")
}
\references{
Hasan and Schliekelman (2017)
}
\author{
Mohamad S. Hasan, \email{shakilmohamad7@gmail.com}
}
