% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_group_df.R
\name{optimal_group_df}
\alias{optimal_group_df}
\title{Obtain optimal number of groups and degrees of freedom}
\usage{
optimal_group_df(group = 5L, pvalue, filter, h_breaks = 71L, m, m1,
  alpha = 0.05, mean_testEffect, effectType = c("continuous", "binary"),
  method = c("BH", "BON"))
}
\arguments{
\item{group}{number of groups to be used to split the p-values, default is five}

\item{pvalue}{a vector of pvalues of the test statistics}

\item{filter}{a vector of filter statistics}

\item{h_breaks}{number of breaks to be used for the histogram, default is 71}

\item{m}{total number of tests}

\item{m1}{number of true alternatve tests}

\item{alpha}{significance level of the hypothesis test}

\item{mean_testEffect}{mean test effect of the true alterantives}

\item{effectType}{type of effect sizes; c("continuous", "binary")}

\item{method}{type of methods is used to obtain the results; c("BH", "BON"),
Benjemini-Hochberg or Bonferroni}
}
\value{
the number of rejected tests and the corresponding Optimal number of
groups and the degrees of freedom of the spline smo0thing
}
\description{
A function to obtain the optimal number of groups and the spline
degrees of freedom from the p-values and the filters by applying the smooting
spline regression
}
\details{
Optimal group and degrees of freedom of the spline regression is a
vital parameter to maximize the number of rejections. This function uses
smooting spline regresion to obtain that.
}
\examples{
# generate pvalues and filter statistics
m = 10000
set.seed(3)
filters = runif(m, min = 0, max = 2.5)          # filter statistics
H = rbinom(m, size = 1, prob = 0.1)             # hypothesis true or false
tests = rnorm(m, mean = H * filters)            # Z-score
pvals = 1 - pnorm(tests)                        # pvalue

results <- optimal_group_df(group = 10, pvalue = pvals, filter = filters,
             h_breaks = 71, m = m, m1 = 8000, alpha = .05,
             mean_testEffect = 2.5, effectType = "continuous", method = "BH")

}
\author{
Mohamad S. Hasan
}
