% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_rank_givenEffect_emp.R
\name{prob_rank_givenEffect_emp}
\alias{prob_rank_givenEffect_emp}
\title{Emperical probbaility of rank of test given effect size}
\usage{
prob_rank_givenEffect_emp(group = 5L, pvalue, filterStat,
  effectType = c("binary", "continuous"))
}
\arguments{
\item{group}{number of groups}

\item{pvalue}{vector of test pvalues}

\item{filterStat}{vector of filter statistics}

\item{effectType}{type of effect size c("binary","continuous")}
}
\value{
\code{prob} emperical probability of the rank of the test
}
\description{
Emperical comnputation of the probbaility of rank of a test being
higher than any other test given the effect size from external information.
}
\details{
If one wants to test \deqn{H_0: \epsilion_i=0 vs. H_a: \epsilion_i > 0,}
then \code{et}  and \code{ey} should be mean of the test and filter effect sizes,
respectively. This is called hypothesis testing for the continuous effect sizes.
If one wants to test \deqn{H_0: \epsilion_i=0 vs. H_a: \epsilion_i = \epsilion,}
then \code{et} and \code{ey} should be median or any discrete value of the
test and filter effect sizes. This is called hypothesis testing for the Binary
effect sizes
}
\examples{

# generating data (known in practice)
pvalue <- runif(100000)
filterStat <- rnorm(100000)

# apply the function to compute the rank proabbility
group=10
ranksProb=sapply(group,prob_rank_givenEffect_emp, pvalue, filterStat,
                       effectType="continuous")

# plot the probability
plot(1:group,ranksProb,type="l",lwd=2,xlab="ranks",ylab="P(rank|effect)")

}
\seealso{
\code{\link{qvalue}} \code{\link{runif}} \code{\link{rnorm}}
}
\author{
Mohamad S. Hasan, mshasan@uga.edu
}
