% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_freq.R
\name{relative_freq}
\alias{relative_freq}
\title{compute relative frequency}
\usage{
relative_freq(bin_idx = 1L, h_breaks = 20L, obs)
}
\arguments{
\item{bin_idx}{Integer, bin number of the histogram}

\item{h_breaks}{number of breaks for the histogram}

\item{obs}{A numeric vector of observations}
}
\value{
\code{rf} A numerical vector of the relative frequencies
 \code{rf_one} A numerical value of the relative frequency
of a particular bin
}
\description{
A function to comoute relative frequency efficiently
}
\details{
This function compute relative frequency effectively using tabulate
}
\examples{

# generating data (known in practice)
X = runif(1000, min = 0, max = 2)         # covariate
r_freq <- relative_freq(bin_idx = 1, h_breaks = 20, obs = X)

}
\author{
Mohamad S. Hasan, mshasan@uga.edu
}
