% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_fwer_emp.R
\name{simu_fwer_emp}
\alias{simu_fwer_emp}
\title{Simulate Family Wise Error Rate (FWER) empirically}
\usage{
simu_fwer_emp(s, m, alphaVec, max.group = 5L)
}
\arguments{
\item{s}{number of replication in a simulation}

\item{m}{total number of hypothesis test}

\item{alphaVec}{a vector of significance levels}

\item{max.group}{maximum number of p-value groups to be used, minimum is 5}
}
\value{
a matrix of fwer for different methods
}
\description{
This function simulate family wise error rate or test type I error rate
}
\details{
This function generate pvalues times form \code{uniform(0, 1)} then applying
OPWeight method to obtain the Familly Wise Error Rate (FWER)
}
\examples{
simVal = 1  # in actual case use at least simVal = 1000
typeIerror_mat = sapply(simVal, simu_fwer_emp, m = 10000, alphaVec = .05)

}
\references{
Hasan and Schliekelman (2017)
}
\seealso{
\code{\link{qvalue}}
\code{\link{prob_rank_givenEffect}}
\code{\link{weight_binary}}
\code{\link{weight_continuous}}
}
\author{
Mohamad S. Hasan, \email{mshasan@uga.edu}
}
