% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empOPW.R
\name{empOPW}
\alias{empOPW}
\title{Perform Empirical Optimal P-value Weighting}
\usage{
empOPW(pvalue, filter, weight = NULL, ranksProb = NULL,
  mean_testEffect = NULL, alpha = 0.05, tail = 1L, delInterval = 0.001,
  group = NULL, max.group = 5L, h_breaks = 71L,
  effectType = c("continuous", "binary"), method = c("BH", "BON"), ...)
}
\arguments{
\item{pvalue}{a vector of pvalues of the test statistics}

\item{filter}{a vector of filter statistics}

\item{weight}{optional weight vector not required}

\item{ranksProb}{ranks probabilities of the test-groups by the filters given
the mean effect. Note that, for each group of tests ranks probbaility would
be the same.}

\item{mean_testEffect}{mean test effect of the true alterantives}

\item{alpha}{significance level of the hypothesis test}

\item{tail}{right-tailed or two-tailed hypothesis test. default is right-tailed test.}

\item{delInterval}{interval between the \code{delta} values of a sequence.
Note that, \code{delta} is a LaGrange multiplier, necessary to normalize the weight}

\item{group}{Integer, number of groups. Default is NULL. If one wants to use
a fixed number of group then should use max.group = NULL}

\item{max.group}{maximum number of groups to be used to split the p-values,
default is five. Note that, it is better to keep approximately 1000 p-values
per group.}

\item{h_breaks}{number of breaks to be used for the histogram, default is 71}

\item{effectType}{type of effect sizes; c("continuous", "binary")}

\item{method}{type of methods is used to obtain the results; c("BH", "BON"),
Benjemini-Hochberg or Bonferroni}

\item{...}{Arguments passed to internal functions}
}
\value{
\code{totalTests} total number of hypothesis tests evaluated

\code{nullProp} estimated propotion of the true null hypothesis

\code{opGroup} Integer, optimal number of groups

\code{ranksProb} probability of the ranks given the mean filter effect,
p(rank | ey = mean_filterEffect)

\code{group_wgt} Numeric vector of group weights (normalized)

\code{weight} Numeric vector of normalized weight for all tests

\code{rejections} total number of rejections

\code{rejections_list} list of rejected pvalues and the corresponding
filter statistics
}
\description{
A function to perform weighted p-value multiple hypothesis test.
This function compute the ranks probability of the test statistics by the filter
statistics given the effect sizes, and consequently the weights if neighter
the weights nor the probabilities are given. Then provides the number of rejected null
hypothesis and the list of the rejected pvalues as well as the corresponing
filter statistics.
}
\details{
If one wants to test \deqn{H_0: epsilon_i = 0 vs. H_a: epsilon_i > 0,}
then the \code{mean_testEffect}  and \code{mean_filterEffect} should be mean
of the test and filter effect sizes, respectively. This is called hypothesis
testing for the continuous effect sizes.\cr

If one wants to test \deqn{H_0: epsilon_i = 0 vs. H_a: epsilon_i = epsilon,}
then \code{mean_testEffect} and \code{mean_filterEffect} should be median or
any discrete value of the test and filter effect sizes. This is called hypothesis
testing for the Binary effect sizes, where \code{epsilon} refers to a fixed value.\cr

The main goal of the function is to compute the probabilities of the ranks from
the pvalues ranked by the filter statistics, consequently the weights.
Although \code{weights} \code{ranksProb} are optional, \code{empOPW} has the
options so that one can compute the probabilities and the weights externally
if necessary (see the examples).\cr

Internally, \code{empOPW} function compute the \code{ranksProb} and consequently
the weights, then uses the p-values to make conclusions about hypotheses.
Although \code{ranksProb} is not required to the function,
One can compute \code{ranksProb} empirically by using the function
\code{\link{prob_rank_givenEffect_emp}}.\cr

The function internally compute \code{mean_testEffect} from the test statistics,
which is obtainde from the p-values.\cr

It is better to see different combinations of groups and h_breaks to optimize
the rejections. The number of p-values per group could be approximately 1000.\cr

One must need to provide either the number of groups or the maximum number of
groups. If both or only the max.group is given then the max.group will be used
to obatian the optimal group;otherwise, the number of groups will be determined
by the group.
}
\examples{
library(OPWeight)
# generate pvalues and filter statistics
m = 10000
set.seed(123)
filters = runif(m, min = 0, max = 2.5)          # filter statistics
H = rbinom(m, size = 1, prob = 0.1)             # hypothesis true or false
tests = rnorm(m, mean = H * filters)            # Z-score
pvals = 1 - pnorm(tests)                        # pvalue

# general use
results <- empOPW(pvalue = pvals, filter = filters, effectType = "continuous",
                                              method = "BH")

# supply the mean test effect externally
library(qvalue)
nullProp = qvalue(p = pvals, pi0.method = "bootstrap")$pi0
m0 = ceiling(nullProp*m)
m1 = m - m0

et = mean(sort(tests, decreasing = TRUE)[1:m1])
results2 <- empOPW(pvalue = pvals, filter = filters, mean_testEffect = et,
               tail = 2, effectType = "continuous", method = "BH")

# supply the ranks probability externally
grp = 5
probs = prob_rank_givenEffect_emp(pvalue = pvals, filter = filters, group = grp,
                               h_breaks = 101, effectType = "continuous")
results3 <- empOPW(pvalue = pvals, filter = filters, ranksProb = probs,
                 effectType = "continuous", tail = 2, method = "BH")

# supply weight externally
wgt <- weight_continuous(alpha = .05, et = et, m = grp, ranksProb = probs)
results4 <- empOPW(pvalue = pvals, filter = filters, weight = wgt,
                        effectType = "continuous", alpha = .05, method = "BH")

}
\seealso{
\code{\link{prob_rank_givenEffect_emp}} \code{\link{weight_binary}}
\code{\link{weight_continuous}}
}
\author{
Mohamad S. Hasan
}
