% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_rank_givenEffect_emp.R
\name{prob_rank_givenEffect_emp}
\alias{prob_rank_givenEffect_emp}
\title{Emperical ranks probbaility of the test given the effect size}
\usage{
prob_rank_givenEffect_emp(pvalue, filter, group = 5L, h_breaks = 100L,
  bin_idx = 1L, smooth = TRUE, effectType = c("continuous", "binary"))
}
\arguments{
\item{pvalue}{vector of test pvalues}

\item{filter}{vector of filter statistics}

\item{group}{number of groups, should be at least four}

\item{h_breaks}{number of breaks for the histogram}

\item{bin_idx}{Integer, bin number of the histogram. Almost alwyas it is one,
becasue we are interested to obtain the ranks probability of the alternative
tests}

\item{smooth}{Character of ("TRUE" or "FALSE") to apply spline smoothing,
default is TRUE}

\item{effectType}{type of effect size c("binary","continuous")}
}
\value{
\code{ranksProb} emperical probability of the rank of the test
}
\description{
Emperical comnputation of the ranks probability of a test being
higher than any other test given the effect size from the external information.
}
\details{
If one wants to test \deqn{H_0: \epsilion_i=0 vs. H_a: \epsilion_i > 0,}
then \code{et}  and \code{ey} should effect type continuous effect size and
if one wants to test \deqn{H_0: \epsilion_i=0 vs. H_a: \epsilion_i = \epsilion,}
one should the binary effect size
}
\examples{

# generating data (known in practice)
set.seed(123)
X = runif(10000, min = 0, max = 2.5)         # covariate
H = rbinom(length(X), size = 1, prob = 0.1)   # hypothesis true or false
Z = rnorm(length(X), mean = H * X)            # Z-score
p = 1 - pnorm(Z)

# apply the function to compute the rank proabbility
grp = 10
ranksProb = prob_rank_givenEffect_emp(pvalue = p, filter = X, group = grp,
                               h_breaks = 71, effectType = "continuous")

# plot the probability
plot(1:grp, ranksProb, type="l", xlab = "ranks", ylab = "P(rank | effect)")

}
\author{
Mohamad S. Hasan, mshasan@uga.edu
}
